requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
pypdf>=4.0.0
chromadb>=0.5.0
openai>=1.40.0
streamlit>=1.38.0
python-dotenv
langchain-text-splitters